Add fatjar generator for AWS Lambdas
AWS Lambda supports 3 methods for uploading a jar. The first two presume
you're shading your jars, which isn't always the best choice. In this
case, I am using the third option, which has non-simple negative impacts
on design.

In this case, I am instead generating a zip file from the jar, and
including it's dependencies in `zip`. This gives you a proper ordered
classpath as expected.

I haven't tested this, outside of checking the input looks loosely how
I'd expect it to. It's possible that the nested lib directory completely
screws this up.