Add main & better CLI parsing to Capsule
Adding main to capsule highlighted too many arguments to the capsule main, it
is overwhelming.  In light of ADR 001, I wanted to take some steps to make the
CLI better too.

Specifying the main simplifies the running of the jar, as it becomes a simple
`java -jar`, with the main baked in.