Accept JAR manifest directly in Capsule API
This is to make the capsule classpath-string->jar function more general.
By accepting a sequence of tuples instead of few enumerated fields it's
possible to extend manifest in a future by additional fields, eg.
related to caplets.

In particular, this makes it easier to use pack's capsule module as an
API.

Fix #5

Signed-off-by: Dominic Monroe <dominic@juxt.pro>