Fix skinny separator colliding with classpath
The ":" character is used for separating elements in the classpath on
Linux, this means that files containing a `:` character are unable to be
added to the classpath.

This uses `__` to separate sections, mostly because it's very visible. A
single _ would perhaps suffice. It isn't the most attractive separator,
and an alternative would be very welcome.