Add system deps.edn by default
This resolves an issue where :mvn/repos is not present by default on
some systems.  This has the drawback of not functioning if the user
changes their system deps.edn or if upstream changes the system
deps.edn.

Finding the real system deps.edn is difficult due to the way
`clojure-env` works, and also because of a desire to avoid the user
deps.edn with a vengeance.

Fix #15
Fix #3