Set Docker working directory
Prior to this change, working directory was unset. This means that that
containerized application cannot rely on relative paths.

I think setting the working directory to the target directory (the
directory where the Clojure application is deployed) is a safe bet.

Best practices for writing Dockerfiles recommends setting the working
directory. See:

https://docs.docker.com/develop/develop-images/dockerfile_best-practices/#workdir